+++ dragino/__init__.py
@@ -1,0 +1,2 @@
+name = "dragino"
+from .dragino import Dragino, DraginoError
